requirements.txt
# Descarga de datasets
kagglehub>=0.2.0

# Procesamiento de datos
pandas>=2.0.0
numpy>=1.24.0

# OCR
pytesseract>=0.3.10
Pillow>=10.0.0

# NLP
nltk>=3.8.0

# Web scraping (opcional, para futuras mejoras)
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Machine Learning tradicional
scikit-learn>=1.3.0

# Visualización
matplotlib>=3.7.0
seaborn>=0.12.0

# Configuración
pyyaml>=6.0

# Progress bars
tqdm>=4.65.0

# Jupyter (para notebooks)
jupyter>=1.0.0
ipykernel>=6.25.0